services/agent-browser/requirements.txt
fastapi==0.115.12
uvicorn==0.32.1
pydantic==2.10.0
python-jobspy==1.1.82
playwright==1.49.0
google-genai==1.5.0
firebase-admin==6.6.0
google-cloud-storage==2.19.0
google-cloud-firestore==2.19.0
